// src/shared/interfaces/book.interface.ts
export interface Book {
  kind: string;
  id: string;
  etag: string;
  selfLink: string;
  volumeInfo: {
    title: string;
    authors: string[];
    publisher: string;
    publishedDate: number;
    description: string;
    industryIdentifiers: {
      type: string;
      identifier: number
    }[];
    readingModes: {
      text: boolean;
      image: boolean
    };
    pageCount: number;
    printType: string;
    categories: string[];
    maturityRating: string;
    allowAnonLogging: boolean;
    contentVersion: string;
    panelizationSummary: {
      containsEpubBubbles: boolean;
      containsImageBubbles: boolean
    };
    imageLinks: {
      smallThumbnail: string;
      thumbnail: string
    };
    language: string;
    previewLink: string;
    infoLink: string;
    canonicalVolumeLink: string
  };
  saleInfo: {
    country: string;
    saleability: string;
    isEbook: boolean
  };
  accessInfo: {
    country: string;
    viewability: string;
    embeddable: boolean;
    publicDomain: boolean;
    textToSpeechPermission: string;
    epub: {
      isAvailable: boolean;
      acsTokenLink: string
    };
    pdf: {
      isAvailable: boolean;
      acsTokenLink: string
    };
    webReaderLink: string;
    accessViewStatus: string;
    quoteSharingAllowed: boolean
  }
}